requirements.txt
# TRCT-GAN Requirements
# Python >= 3.8

# Core Deep Learning
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.21.0
scipy>=1.7.0

# Data Processing
nibabel>=3.2.0
SimpleITK>=2.1.0
opencv-python>=4.5.0
Pillow>=8.3.0
scikit-image>=0.19.0

# Configuration and Utilities
pyyaml>=5.4.0
tqdm>=4.62.0
einops>=0.4.0

# Logging and Visualization
tensorboard>=2.10.0
matplotlib>=3.4.0
seaborn>=0.11.0

# Optional: Weights & Biases
# wandb>=0.12.0

# Optional: DICOM support
# pydicom>=2.2.0

# Development
pytest>=6.2.0
black>=21.9b0
flake8>=3.9.0
